moved generateRefreshCookie method to CookieService class
